feat(navbar): Refactor structure and Products dropdown menu
